refactoring with relocating data to index.js, add 2 new components to dialogs.jsx, present bag in message people
