+++ src/components/Dialogs/DialogItem/DialogItem.jsx
@@ -1,0 +1,15 @@
+import s from "../Dialogs.module.css";
+import React from "react";
+import { NavLink } from "react-router-dom";
+
+const DialogItem = (props) => {
+  let path = "/dialogs/" + props.id;
+  return (
+    <div className={s.dialog + " " + s.active}>
+      <NavLink to={path}>{props.name}</NavLink>
+    </div>
+  );
+};
+
+
+export default DialogItem;
