+++ src/components/Dialogs/Dialogs.jsx
@@ -1,45 +1,14 @@
 import s from "./Dialogs.module.css";
 import React from "react";
-import { NavLink } from "react-router-dom";
-
-const DialogItem = (props) => {
-  let path = "/dialogs/" + props.id;
-  return (
-    <div className={s.dialog + " " + s.active}>
-      <NavLink to={path}>{props.name}</NavLink>
-    </div>
-  );
-};
-
-const Message = (props) => {
-    return (
-        <div className={s.message}>{props.message}</div>
-    )
-}
+import DialogItem from './DialogItem/DialogItem'
+import Message from "./Message/Message";
 
 const Dialogs = (props) => {
-  
-    let dialogsData = [
-      {id: 1, name: 'Andrii'},
-      {id: 2, name: 'Dima'},
-      {id: 3, name: 'Marik'},
-      {id: 4, name: 'Vova'},
-      {id: 5, name: 'Roman'},
-      {id: 6, name: 'Bohdan'},
-  ];
 
-  let dialogsElements = dialogsData.map( elem => <DialogItem name={elem.name} id={elem.id}/>)
- 
-  let messagesData = [
-    {id: 1, message: 'Hi'},
-    {id: 2, message: 'How is your dog?'},
-    {id: 3, message: 'Yo'},
-    {id: 4, message: 'Hello'},
-    {id: 5, message: 'Bye'},
-    {id: 6, message: 'QQ'},
-] 
+  let dialogsElements = props.dialogsData
+    .map( elem => <DialogItem name={elem.name} id={elem.id}/>)
 
-  let messagesElements = messagesData
+  let messagesElements = props.messagesData
     .map( message => <Message message={message.message} id={message.message} />)
 
   return (
