+++ src/components/Profile/Profile.jsx
@@ -3,11 +3,12 @@ import MyPosts from "./MyPosts/MyPosts";
 import s from "./Profile.module.css";
 import ProfileInfo from "./ProfileInfo/ProfileInfo";
 
-const Profile = () => {
+const Profile = (props) => {
+  
   return (
     <div className={s.profile}>
       <ProfileInfo/>
-      <MyPosts/>
+      <MyPosts posts={props.posts}/>
     </div>
   );
 };
